requirements.txt
opencv-python-headless
mediapipe
numpy
Pillow
ipywidgets
matplotlib
